backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database and TimescaleDB
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Time series forecasting (REAL models)
statsmodels==0.14.0
prophet==1.1.4
scikit-learn==1.3.2
torch==2.1.1
pytorch-lightning==2.1.2

# Financial data APIs (REAL data sources)
yfinance==0.2.28
requests==2.31.0
aiohttp==3.9.1

# Risk metrics and portfolio analysis
pyfolio==0.9.2
empyrical==0.5.5
QuantLib==1.32

# Explainability and AI
shap==0.44.0
openai==1.3.7
tiktoken==0.5.2

# Visualization and charts
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production monitoring
prometheus-client==0.19.0